// packages/engine/src/input/components/XRControllersComponent.d.ts
import { Component } from '../../ecs/classes/Component';
export declare class XRControllersComponent extends Component<any> {
    static _schema: {
        position: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        controller: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        controller1: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        controller2: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        position1: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        position2: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        rotation1: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        rotation2: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        controllerGrip1: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        controllerGrip2: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        physicsBody1: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
        physicsBody2: {
            type: import("@xr3ngine/engine/src/ecs/types/Types").PropType<unknown, any>;
            default: any;
        };
    };
}